[BUG] Map-Creator: Fix sync of some datas

+++ Server/Handler/Entities/LobbySystem/MapCreateLobby/Sync.cs
@@ -81,7 +81,7 @@ namespace TDS_Server.Handler.Entities.LobbySystem
 
         public void SyncMapInfoChange(MapCreatorInfoType infoType, object data)
         {
-            ModAPI.Sync.SendEvent(this, ToClientEvent.ToBrowserEvent, ToBrowserEvent.MapCreatorSyncData, infoType, data);
+            ModAPI.Sync.SendEvent(this, ToClientEvent.ToBrowserEvent, ToBrowserEvent.MapCreatorSyncData, (int)infoType, data);
 
             switch (infoType)
             {
